Port rust-lang/rust documentation changes
The following are some of the PRs included here:

- #92902: Improve the documentation of drain members
- #87537: Clarify undefined behaviour in binary heap, btree and hashset
  docs
- #89010: Add some intra doc links
- #80681: Clarify what the effects of a 'logic error' are
- #77079: Use Self in docs when possible
- #75974: Avoid useless sift_down when
  std::collections::binary_heap::PeekMut is never mutably dereferenced
- #76534: Add doc comments for From impls
- #75831: doc: Prefer https link for wikipedia URLs
- #74010: Use italics for O notation
- #71167: big-O notation: parenthesis for function calls, explicit
  multiplication
- #63486: Document From trait for BinaryHeap
- #60952: Document BinaryHeap time complexity
- #60451: BinaryHeap: add min-heap example

Also port the change in rust-lang/rust@99ed06eb88 "libs: doc comments".

Note that rust-lang/rust#60451 adds an example of a min-heap.  We add a
similar example here, although edited to highlight
`binary_heap_plus::BinaryHeap`'s ability to implement a min-heap
*without* wrapping the items in `std::cmp::Reverse`.

Finally, we replace the wildcard import in the documentation examples
with a named import of `BinaryHeap`.