Add additional test from `alloc`
This test is taken from
`library/alloc/src/collections/binary_heap/tests.rs` in Rust 1.62.0.

Note that the setup in the test is adapted to this crate by implementing
`Ord` instead of `PartialOrd` for `PanicOrd<T>` as is done upstream,
since the heap operations here rely on the `Ord` trait whereas upstream
they rely on the `PartialOrd` trait.