Port updated example from rust-lang/rust
This commits ports one small part of rust-lang/rust#91861.  The main
change that PR made to `binary_heap.rs` was to replace vectors with
arrays in the examples.  We do *not* port such changes as they require
Rust 1.56.0, which is greater than this crate's current MSRV.  However,
it also simplified the setup in the example of `BinaryHeap::append()`,
and we repeat that here only with vectors instead of arrays.